HBX-1605: Move class 'org.hibernate.cfg.binder.PrimaryKeyInfo' to 'org.hibernate.tool.internal.reveng.PrimaryKeyInfo'
